+++ src/ravn/adapters/tools/web_fetch.py
@@ -2,15 +2,14 @@
 
 from __future__ import annotations
 
-import ipaddress
 import logging
 import re
-import socket
 from html.parser import HTMLParser
 from urllib.parse import urlparse
 
 import httpx
 
+from ravn.adapters.tools._url_security import check_ssrf
 from ravn.domain.models import ToolResult
 from ravn.ports.tool import ToolPort
 
@@ -106,25 +105,6 @@ def truncate_to_budget(text: str, budget: int) -> str:
 # SSRF protection
 # ---------------------------------------------------------------------------
 
-_PRIVATE_NETWORKS: list[ipaddress.IPv4Network | ipaddress.IPv6Network] = [
-    ipaddress.ip_network("127.0.0.0/8"),
-    ipaddress.ip_network("10.0.0.0/8"),
-    ipaddress.ip_network("172.16.0.0/12"),
-    ipaddress.ip_network("192.168.0.0/16"),
-    ipaddress.ip_network("169.254.0.0/16"),
-    ipaddress.ip_network("::1/128"),
-    ipaddress.ip_network("fc00::/7"),
-]
-
-
-def _is_private_ip(ip: str) -> bool:
-    """Return True if the IP address falls within a private/reserved range."""
-    try:
-        addr = ipaddress.ip_address(ip)
-    except ValueError:
-        return True  # Unparseable — block by default.
-    return any(addr in net for net in _PRIVATE_NETWORKS)
-
 
 def _validate_url(url: str) -> str | None:
     """Return an error message if *url* is disallowed, else None.
@@ -141,17 +121,7 @@ def _validate_url(url: str) -> str | None:
     if not hostname:
         return "Blocked: URL has no hostname"
 
-    try:
-        results = socket.getaddrinfo(hostname, None)
-    except OSError:
-        return f"Blocked: could not resolve hostname '{hostname}'"
-
-    for _family, _type, _proto, _canonname, sockaddr in results:
-        ip = sockaddr[0]
-        if _is_private_ip(ip):
-            return f"Blocked: '{hostname}' resolves to a private/reserved address"
-
-    return None
+    return check_ssrf(hostname)
 
 
 # ---------------------------------------------------------------------------
